bpf: Factor out the element allocation for pre-allocated htab
The element allocation for pre-allocated htab is composed of two logics:
1) when there is old_element, directly reuse per-cpu extra_elems
Also stash the old_element as the next per-cpu extra_elems
2) when no old_element, allocate from per-cpu free list

The reuse and stash of per-cpu extra_elems will be broken into two
independent steps. After the breaking, per-cpu extra_elems may be NULL
when trying to reuse it and the allocation needs to fall-back to per-cpu
free list when it happens.

Therefore, factor out the element allocation to a helper to make the
following change be straightforward.

Signed-off-by: Hou Tao <houtao1@huawei.com>